requirements.txt
amqp==2.6.1
aniso8601==3.0.2
asn1crypto==0.24.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.11.5
chardet==4.0.0
Click==7.0
cryptography==2.3.1
Flask==1.0.2
Flask-Caching==1.9.0
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
flower==0.9.7
humanize==2.6.0
idna==2.7
importlib-metadata==2.1.1
ItsDangerous==1.0.0
Jinja2==2.10
kombu==4.6.11
MarkupSafe==1.0
pkg-resources==0.0.0
prometheus-client==0.8.0
pycparser==2.19
python-dotenv
PyMySQL==0.9.2
pytz==2018.5
redis==3.5.3
requests==2.25.1
six==1.11.0
tornado==6.1
urllib3==1.26.3
uWSGI==2.0.17.1
vine==1.3.0
Werkzeug==0.14.1
zipp==1.2.0

